requirements.txt
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/
--extra-index-url=http://dist.pinaxproject.com/fresh-start/

Django==1.3.1
Markdown==2.1.0
Pinax==0.9a2
django-announcements==0.1.2
django-appconf==0.4.1
django-compressor==1.0.1
django-debug-toolbar==0.8.5
django-email-confirmation==0.2
django-mailer==0.2a1.dev3
django-notification==0.2
django-openid==0.3a1
django-pagination==1.0.7
django-staticfiles==1.1.2
django-taggit==0.9.3
django-taggit-templatetags==0.4.6dev
django-templatetag-sugar==0.1
django-timezones==0.2
gunicorn==0.13.4
idios==0.2a1
metron==0.1
pinax-theme-bootstrap==0.1.2
psycopg2==2.4.3
python-openid==2.2.5
pytz==2011n
wsgiref==0.1.2

git+git://github.com/toastdriven/django-haystack.git@#egg=django-haystack-2.0.0-beta

Whoosh==2.3.2
argparse==1.2.1
haystack==0.13
python-ptrace==0.6.3
django-tastypie==0.9.11
mimeparse==0.1.3
python-dateutil==1.5